azure-vision-project/requirement.txt
azure-cognitiveservices-vision-computervision
opencv-python
numpy
